Implement nodes.py module.
This module contains a single function, tree(). The tree() function
examines the loggers currently registered with Python in the
`loggerDict` dictionary, which resides on the logging manager, a global
object. The function enumerates the items in the dictionary and
constructs a tree representing all the loggers.

Each node in the tree is a three tuple.